requirements.txt
fastapi
uvicorn
tensorflow-cpu
scikit-learn
joblib
numpy
python-multipart
opencv-python-headless
mediapipe
